Python Script to generate queries per thread.
This python script will generate the specified number of queries per thread. The first argument is the base file and can be found in the folder supporting files. The of the base file is
For lock free: "LockFreeList_4Threads.cpp"
For lock: "ListUsingLocks_4_threads.cpp"

Currenty the script expects the base code to have 4 threads.